add struct files
